requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-bootstrap==0.2.4
django-bootstrap4==2.3.1
django-classy-tags==2.0.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-datetimepicker==3.14
django-heroku==0.3.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
httpie==2.0.0
idna==2.8
inflection==0.3.1
ipaddr==2.2.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
pep517==0.8.2
Pillow==7.0.0
pipenv==2020.11.15
progress==1.5
psycopg2==2.8.6
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.11.0
PyMySQL==0.10.1
pyparsing==2.4.6
python-decouple==3.3
python3-openid==3.1.0
pytoml==0.1.21
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==20.4.2
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.11.2
